+++ src/spec/lib/NsidcOpenSearchResponse_spec.js
@@ -1,4 +1,5 @@
 import NsidcOpenSearchResponse from '../../lib/NsidcOpenSearchResponse';
+import _ from "underscore";
 
 describe('NsidcOpenSearchResponse', function () {
     let nsidcOpenSearchResponse = new NsidcOpenSearchResponse();
@@ -6,6 +7,7 @@ describe('NsidcOpenSearchResponse', function () {
     let fakeResponseHeader = '<feed xmlns="http://www.w3.org/2005/Atom"' +
       ' xmlns:os="http://a9.com/-/spec/opensearch/1.1/"' +
       ' xmlns:dif="http://gcmd.gsfc.nasa.gov/Aboutus/xml/dif/"' +
+      ' xmlns:dc="http://purl.org/dc/elements/1.1/"' +
       ' xmlns:nsidc="http://nsidc.org/ns/opensearch/1.1/">' +
       '<title>NSIDC dataset search results</title>' +
       '<os:totalResults>1</os:totalResults>' +
@@ -18,11 +20,19 @@ describe('NsidcOpenSearchResponse', function () {
       '<temporal_duration>12345</temporal_duration>' +
       '<spatial_area>180.0</spatial_area>' +
       '<summary>dataset summary</summary>' +
+      '<dc:date>1970-01-01/2000-12-31</dc:date>' +
       '<author><name>author one</name></author>' +
       '<dif:Keyword>keyword one</dif:Keyword>' +
       '<dif:Parameters><dif:Category>some category</dif:Category><dif:Topic>some topic</dif:Topic><dif:Term>some term</dif:Term>' +
       '<dif:Detailed_Variable>variable details</dif:Detailed_Variable></dif:Parameters>' +
       '<dif:Distribution><dif:Distribution_Format>stone tablet</dif:Distribution_Format></dif:Distribution>';
+
+    // Reformat an array of date strings into a single XML string
+    function dateRangesToXml(rangeList) {
+        return _.map(rangeList, function(range) {
+            return '<dc:date>' + range.join('/') + '</dc:date>';
+        }).join(' ');
+    }
 
     it('has a creation method from xml', function () {
         expect(nsidcOpenSearchResponse.fromXml instanceof Function).toBe(true);
@@ -37,7 +47,7 @@ describe('NsidcOpenSearchResponse', function () {
     describe('Parse XML response', function () {
         it('extracts OS fields from response', function () {
             let fakeResponse = fakeResponseHeader + '<entry>' + fakeResponseEntry + '</entry></feed>';
-            var results = nsidcOpenSearchResponse.fromXml(fakeResponse, {osSearchTerms: '', osDtStart: '', osDtEnd: '', geoBoundingBox: ''});
+            let results = nsidcOpenSearchResponse.fromXml(fakeResponse, {});
 
             expect(results.getTotalCount()).toEqual(1);
             expect(results.getCurrentIndex()).toEqual(1);
@@ -49,10 +59,68 @@ describe('NsidcOpenSearchResponse', function () {
             expect(results.getResults()[0].keywords).toEqual(['keyword one']);
             expect(results.getResults()[0].parameters).toEqual(['variable details']);
             expect(results.getResults()[0].dataFormats).toEqual(['stone tablet']);
+            expect(results.getResults()[0].dateRanges).toEqual([{startDate: '1970-01-01', endDate: '2000-12-31'}]);
         });
 
-        // A few tests were updated as part of SRCH-94. The skipped tests still need to be
-        // updated to use updated test data and to run in the Jest framework.
+        it('extracts multiple complete date ranges', function() {
+            let rangeList = [
+              ['1980-01-01', '2010-12-31'],
+              ['1990-01-01', '2020-12-31']
+            ];
+            let moreFakeRanges = dateRangesToXml(rangeList);
+
+            let fakeResponse = fakeResponseHeader + '<entry>' +
+              fakeResponseEntry + moreFakeRanges + '</entry></feed>';
+
+            let results = nsidcOpenSearchResponse.fromXml(fakeResponse, {});
+            let dateRanges = results.getResults()[0].dateRanges;
+            expect(dateRanges).toBeInstanceOf(Array);
+            expect(dateRanges.length).toEqual(3);
+            expect(dateRanges[0].startDate).toBe('1970-01-01');
+            expect(dateRanges[2].endDate).toBe('2020-12-31');
+        });
+
+        it('extracts date ranges with empty values', function() {
+            let rangeList = [
+                ['1980-01-01', ''],
+                ['', '2020-12-31'],
+                ['', '']
+            ];
+            let moreFakeRanges = dateRangesToXml(rangeList);
+            let fakeResponse = fakeResponseHeader + '<entry>' +
+              fakeResponseEntry + moreFakeRanges + '</entry></feed>';
+
+            let results = nsidcOpenSearchResponse.fromXml(fakeResponse, {});
+            let dateRanges = results.getResults()[0].dateRanges;
+
+            expect(dateRanges.length).toEqual(4);
+            expect(dateRanges[1].startDate).toBe('1980-01-01');
+            expect(dateRanges[1].endDate).toBe('');
+            expect(dateRanges[2].startDate).toBe('');
+            expect(dateRanges[2].endDate).toBe('2020-12-31');
+            expect(dateRanges[3].startDate).toBe('');
+            expect(dateRanges[3].endDate).toBe('');
+        });
+
+        it('ignores date ranges with no separator character', function() {
+            let fakeResponse = fakeResponseHeader + '<entry>' +
+              fakeResponseEntry + '<dc:date></dc:date>' + '</entry></feed>';
+            let results = nsidcOpenSearchResponse.fromXml(fakeResponse, {});
+            let dateRanges = results.getResults()[0].dateRanges;
+            expect(dateRanges.length).toEqual(1);
+            expect(dateRanges[0].startDate).toBe('1970-01-01');
+
+            fakeResponse = fakeResponseHeader + '<entry>' +
+              fakeResponseEntry + '<dc:date>nonsense</dc:date>' + '</entry></feed>';
+            results = nsidcOpenSearchResponse.fromXml(fakeResponse, {});
+            dateRanges = results.getResults()[0].dateRanges;
+            expect(dateRanges.length).toEqual(1);
+            expect(dateRanges[0].startDate).toBe('1970-01-01');
+        });
+
+        // A few tests were updated or added as part of SRCH-94 and SRCH-51. The
+        // skipped tests still need to be updated to use updated test data and to
+        // run in the Jest framework.
         it.skip('extracts multiple data access links', function () {
             var results, fakeFeed, additionalEntryContents = [
                 '   <link title="' + fakeDataUrls[1].title + '" href="' + fakeDataUrls[1].href +